reorganized files a bit
